# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baskonia Pipeline API",
    "description": "API de análisis de Baskonia (fase F3 de la migración strangler-fig).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{slug}/games": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List Games",
        "description": "Partidos de un equipo (jugados y pendientes), paginados (endpoint 7).",
        "operationId": "list_games_api_v1_teams__slug__games_get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo (p.ej. 'vitoria')",
              "title": "Slug"
            },
            "description": "Slug del equipo (p.ej. 'vitoria')"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "season",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Año de inicio de temporada (p.ej. 2025)",
              "title": "Season"
            },
            "description": "Año de inicio de temporada (p.ej. 2025)"
          },
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Competición: acb | euroleague | supercopa",
              "title": "League"
            },
            "description": "Competición: acb | euroleague | supercopa"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GamesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/games/{game_id}/boxscore": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Get Boxscore",
        "description": "Box score de un equipo en un partido (endpoint 16).\n\n`team_slug` es un query param obligatorio (no va en la ruta).",
        "operationId": "get_boxscore_api_v1_games__game_id__boxscore_get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          },
          {
            "name": "team_slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Slug del equipo cuyo box score se pide",
              "title": "Team Slug"
            },
            "description": "Slug del equipo cuyo box score se pide"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoxScoreResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoxScoreResponse": {
        "properties": {
          "game_id": {
            "type": "integer",
            "title": "Game Id"
          },
          "team": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "opponent": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "team_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Score"
          },
          "opponent_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Score"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/BoxScoreRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "team",
          "opponent",
          "date",
          "league",
          "rows"
        ],
        "title": "BoxScoreResponse",
        "description": "Box score de un equipo en un partido."
      },
      "BoxScoreRow": {
        "properties": {
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "rebounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rebounds"
          },
          "assists": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assists"
          },
          "steals": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steals"
          },
          "blocks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocks"
          },
          "turnovers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnovers"
          },
          "fg_made": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fg Made"
          },
          "fg_attempted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fg Attempted"
          },
          "fg3_made": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fg3 Made"
          },
          "fg3_attempted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fg3 Attempted"
          },
          "ft_made": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ft Made"
          },
          "ft_attempted": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ft Attempted"
          },
          "efg_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Efg Pct"
          },
          "ts_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts Pct"
          }
        },
        "type": "object",
        "required": [
          "player_name"
        ],
        "title": "BoxScoreRow",
        "description": "Fila de box score de un jugador en un partido."
      },
...
      "GameAdvanced": {
        "properties": {
          "pace": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pace"
          },
          "off_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Off Rating"
          },
          "def_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Def Rating"
          },
          "net_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Rating"
          }
        },
        "type": "object",
        "title": "GameAdvanced",
        "description": "Estadísticas avanzadas de un equipo en un partido."
      },
      "GameItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "is_home": {
            "type": "boolean",
            "title": "Is Home"
          },
          "opponent": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "team_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Score"
          },
          "opponent_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Score"
          },
          "result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "advanced": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameAdvanced"
              },
              {
                "type": "null"
              }
            ]
          },
          "has_boxscore": {
            "type": "boolean",
            "title": "Has Boxscore",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "league",
          "is_home",
          "opponent"
        ],
        "title": "GameItem",
        "description": "Un partido de un equipo (jugado o pendiente)."
      },
      "GamesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "GamesResponse",
        "description": "Lista paginada de partidos de un equipo."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TeamRef": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name"
        ],
        "title": "TeamRef",
        "description": "Referencia a un equipo (slug + nombre para mostrar)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}